Move Android Application to android subpackge
